Moved test from time series tests to explainer test.
